Emit log messages if an uresolved link was found in the HTML output.
Very naive, just looking for ']['. Will clean up depending on how many
false positives we get.

+++ _scripts/cfdoc_qa.py
@@ -24,8 +24,12 @@ import os
 import sys
 from time import gmtime, strftime
 
-def initialize(config):
+def initialize(config, wipe):
 	config["log_file"] = config["markdown_directory"] + "/cfdoc_log.markdown"
+	
+	if not wipe:
+		return
+
 	if os.path.exists(config["log_file"]):
 		os.remove(config["log_file"])
 	logfile = open(config["log_file"], "w")
